Revert "pinctrl: renesas: Support RGID"
This reverts commit 5021d7e8d8802a41d4f8e4c0ab8a6a3cf7252fce.

+++ drivers/pinctrl/renesas/pfc-r8a77995.c
@@ -2822,7 +2822,7 @@ static const struct pinmux_cfg_reg pinmux_config_regs[] = {
 	{ },
 };
 
-static int r8a77995_pin_to_pocctrl(struct sh_pfc *pfc, unsigned int pin, u64 *pocctrl)
+static int r8a77995_pin_to_pocctrl(struct sh_pfc *pfc, unsigned int pin, u32 *pocctrl)
 {
 	int bit = -EINVAL;
